Spilt DAG and Def to 2 file
